requirements.txt
numpy==1.26.4
pandas==2.2.1
streamlit==1.31.1
pillow==10.3.0